requirements.txt
Django==2.0.9
django-allauth==0.36.0
django-appconf==1.0.2
django-autofixture==0.12.1
django-cors-headers==1.2.2
django-crispy-forms==1.7.2
django-debug-toolbar==1.10.1
django-extensions==2.1.0
django-filter==2.0.0
django-image-cropping==1.2.0
django-js-asset==1.1.0
django-mptt==0.9.1
django-reversion==3.0.2
django-taggit==0.23.0
django-markdownx==2.0.25
djangorestframework==3.8.2
easy-thumbnails==2.5
Markdown==3.0.1
MarkupSafe==1.1.0
oauthlib==2.0.7
Pillow==5.3.0
psycopg2==2.7.5
#psycopg2-binary==2.7.5
PyGithub==1.43.3
PyJWT==1.6.4
python-memcached==1.59
python3-openid==3.1.0
pytz==2018.3
requests==2.20
requests-oauthlib==1.0
six==1.11.0
sqlparse==0.2.4
whitenoise==3.3.1
django-tagulous==0.13.2
